Refactor: 코멘트 api 리팩토링 - S11P21S002-258

+++ frontend/src/components/ImageCanvas/CommentLabel.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import { Group, Rect, Text, Image } from 'react-konva';
 import { CommentResponse } from '@/types';
 import useImage from 'use-image';
@@ -9,13 +9,22 @@ import toggleDownIconSrc from '@/assets/icons/chevron-down.svg';
 import Konva from 'konva';
 
 interface CommentLabelProps {
+  stage: Konva.Stage;
   comment: CommentResponse & { isOpen?: boolean };
   updateComment: (comment: CommentResponse) => void;
   deleteComment: (commentId: number) => void;
   toggleComment: (commentId: number) => void;
 }
 
-export default function CommentLabel({ comment, updateComment, deleteComment, toggleComment }: CommentLabelProps) {
+export default function CommentLabel({
+  stage,
+  comment,
+  updateComment,
+  deleteComment,
+  toggleComment,
+}: CommentLabelProps) {
+  const groupRef = useRef<Konva.Group>(null);
+  // const stage = groupRef.current?.getStage();
   const [content, setContent] = useState(comment.content);
   const [deleteIcon] = useImage(deleteIconSrc);
   const [toggleUpIcon] = useImage(toggleUpIconSrc);
@@ -31,7 +40,7 @@ export default function CommentLabel({ comment, updateComment, deleteComment, to
 
   const handleDelete = (e: Konva.KonvaEventObject<MouseEvent>) => {
     e.cancelBubble = true;
-    deleteComment(comment.id);
+    confirm('정말 삭제하시겠습니까?') && deleteComment(comment.id);
   };
 
   const handleToggle = (e: Konva.KonvaEventObject<MouseEvent>) => {
@@ -40,66 +49,58 @@ export default function CommentLabel({ comment, updateComment, deleteComment, to
   };
 
   return (
-    <Group
-      x={comment.positionX}
-      y={comment.positionY}
-      draggable
-      onDragEnd={(e) => {
-        const newX = e.target.x();
-        const newY = e.target.y();
-        updateComment({ ...comment, positionX: newX, positionY: newY });
-      }}
-    >
-      <Rect
-        width={comment.isOpen ? 200 : 50}
-        height={comment.isOpen ? 100 : 30}
-        fill="white"
-        stroke="black"
-        onClick={handleEdit}
-      />
-      {comment.isOpen && (
-        <Text
+    stage && (
+      <Group
+        ref={groupRef}
+        x={comment.positionX}
+        y={comment.positionY}
+        draggable
+        onDragEnd={(e) => {
+          const newX = e.target.x();
+          const newY = e.target.y();
+          updateComment({ ...comment, positionX: newX, positionY: newY });
+        }}
+        strokeScaleEnabled={false}
+        scale={{ x: 1 / stage.getAbsoluteScale().x, y: 1 / stage.getAbsoluteScale().y }}
+        perfectDrawEnabled={false}
+        shadowForStrokeEnabled={false}
+      >
+        <Rect
+          width={comment.isOpen ? 200 : 60}
+          height={comment.isOpen ? 100 : 30}
+          fill="white"
+          stroke="#080808"
+          strokeWidth={1}
+          cornerRadius={5}
+        />
+        {comment.isOpen && (
+          <Text
+            x={10}
+            y={35}
+            width={190}
+            text={content || '내용 없음'}
+            fontSize={16}
+            fill="#080808"
+            onClick={handleEdit}
+          />
+        )}
+        <Image
+          image={comment.isOpen ? toggleUpIcon : toggleDownIcon}
           x={5}
           y={5}
-          width={190}
-          text={content || '내용 없음'}
-          fontSize={16}
-          fill="black"
-          onClick={handleEdit}
+          width={20}
+          height={20}
+          onClick={handleToggle}
         />
-      )}
-      {deleteIcon && (
         <Image
           image={deleteIcon}
-          x={comment.isOpen ? 175 : 25}
+          x={35}
           y={5}
           width={20}
           height={20}
           onClick={handleDelete}
         />
-      )}
-
-      {comment.isOpen
-        ? toggleUpIcon && (
-            <Image
-              image={toggleUpIcon}
-              x={comment.isOpen ? 150 : 0}
-              y={5}
-              width={20}
-              height={20}
-              onClick={handleToggle}
-            />
-          )
-        : toggleDownIcon && (
-            <Image
-              image={toggleDownIcon}
-              x={comment.isOpen ? 150 : 0}
-              y={5}
-              width={20}
-              height={20}
-              onClick={handleToggle}
-            />
-          )}
-    </Group>
+      </Group>
+    )
   );
 }
